sql: cherry-pick src/js/internal/sql/ BasePooledConnection extraction from #31994
Baseline cherry-pick of origin/claude/split/sql:src/js/internal/sql/.
This intentionally clobbers #32041 and #32097 — they get re-integrated
into BasePooledConnection in the next commit(s).
